+++ src/solvers/SolverBase.hpp
@@ -12,10 +12,10 @@ public:
 protected:
     virtual void setup_system() = 0;
     virtual void assemble_system() = 0;
-    virtual size_t solve_linear_system() = 0;
+    virtual unsigned int solve_linear_system() = 0;
     virtual void do_postprocessing() = 0;
 
-    virtual size_t get_n_dofs() = 0;
+    virtual unsigned int get_n_dofs() = 0;
 };
 
 }
